Move start end date for tab to DateTimeUtils
